TIKA-4789 -- drop solr 8, add 10
